document: make work with utf8 correct
Rewrite some functions in the document.go to
work correct with utf8 characters.

document.go follow next notation:
1) `cursor` means index in the rune array.
2) `index`  means index in the byte array.

Some functions violated this contract.

`findLineStartIndex` must take an index as an argument.
(Because, this functions performs subtraction of an argument
with some other index).
However, functions `CursorPositionRow`, `CursorPositionCol`, passed
the cursor instead of the index.
To fix this, a function that translates the current cursor to an index
was written (`getCursorIndex`).
Also,`GetCursorPosition`, `GetCustomCursorPosition`
were added for convenience.

`GetCursorLeftPosition` operated with `len(string)` instead of
`len([]rune)`. Fixed it.
`GetCursorRightPosition` fixed automatically, because it used another
fixed function inside.

`GetCursorUpPosition` and `GetCursorDownPosition` contained similar
mistakes.
To fix this `TranslateRowColToIndex` was reworked to
`TranslateRowColToCursor`.

The tests have been expanded with `utf8` cases.