render: improve API, add support for multi-line commands
Firstly, this patch bring some logical changes.

At the moment, some attributes, such `prefix`,
`livePrefixCallback`, `breakLineCallback, etc, contained in
the `Render` struct in the `render.go`. However, these attributes
not directly related with rendering itself and represent parts of
the prompt. So, these attributes were moved to struct
`Prompt`.

Besides, working with `Render` function difficult and unclear at
the moment. This function prints something to the out and changes
internal state of the `Prompt` (`prevCursor` attribute) and it's
unclear for reader, where it is possible to call this function in
`prompt.go`, and where it's not. So, logic of this function was changed.
This function performs console output and returns relative cursor
position and we can mantain this values from `prompt.go` and
make state changes manually (it will be useful at the moment of
splitting library into packages in the future).

Also, existence of many arguments makes this function
uncomfortable and difficult to expand.
So, `renderCtx` structure was introduced.
Also, at the moment there are few rendering functions in
`render.go`, which called from `prompt.go` and this pretty
inconvenient. So, introduced main `Render` function that
calls the right sub-render function in dependency of render-event,
which it extracts from the context.

Current rendering pipeline is:
1) prompt forms rendering query using `getCmdToRender` function
2) prompt send this query to render, which can process such
queries, using `writeCmdWithPrefix`, `renderCtx` functions.

The changes above will allow developers to add new features much
simplier.

Secondly, because of rendering was reworked to work in general,
this patch automatically solves the problem with multi-line
commands.