test: add integration tests
Added integration tests.
The tests use `tmux` as a terminal emulator.

To test the library the sample prompt_app has been added.

Also, a mechanism for rendering notifications has been
added. It is used in integration tests to wait terminal
redraws. The mechanism is implemented through Unix sockets.
The testing process starts a notification server and passes
uri of the socket to the test prompt application using the
corresponding `Option`.

In addition, an pytest entry in .gitignore has been added.