prompt: add reverse-search
A reverse-search function has been added to the prompt.

First, added a structure `reverseSearchState`, which
stores information about the current reverse-search session.

Reverse-search pipeline looks as follows:
1) The user starts a `reverse-search` using `Ctrl-R`
2) `Prompt` creates new `reverseSearchState` object, which
refers to the it's history
3) When prompt input is updated, `update` is called for `reverseSearch`
attribute to update the matched command
4) The user can trigger `reverse-search` again using `Ctrl-R`,
after which the search area will be reduced
5) When reverse-search mode is turned off, the state object is destroyed

The Reverse-search rendering was implemented in a natural way: since
rendering of the reverse-search is the same as for the simple commands.
It also has the prefix (with user input) and an active command (matched command).
So, adding this feature affected `getCurrentPrefix` and
`getCmdToRender` functions in `prompt.go`.