render: fix blinking
In some terminals (e.g, `urxvt`) escape-sequence, which is used to
erase the bottom of the console causes commands to blink.

Therefore, the use of this sequence has been replaced by the use of a
sequence, that softly erases line by line.
But, when resizing the window, we cannot move the cursor up line by line,
as this causes prompt to float.

So a new render event was introduced to intercept the window resizing
and use the old erase in this case.