prompt: fix some keybindings to work with multilines
- The behavior of the `Home`/`End` has been changed:
now, the cursor moves to the beginning/end of the entire command
(instead of the line) when these keys are pressed.

- The behavior of the `alt-b`/`alt-f` has been changed:
now, the cursor moves one word backward in multiline commands, and
the word delimeter characters are `\n` and ` `.

- The associated emacs keybindings have been changed:
now `ctrl-b`/`ctrl-f` allow to move over the entire multiline command.
`ctrl-a`/`ctrl-e` same as `Home`/`End`.

- Added integration tests for emacs keyset.

Closes #6